requirements.txt
pybit
pandas
numpy
ta
requests
pyyaml
apscheduler
